requirements.txt
# Core ML & Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.1
lightgbm==4.0.0
optuna==3.14.0

# Time Series & Forecasting
statsmodels==0.14.0
prophet==1.1.5
tensorflow==2.13.0
keras==2.13.0

# Explainability & Interpretability
shap==0.42.1
lime==0.2.0
matplotlib==3.7.2
seaborn==0.12.2

# API & Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Data Drift & Monitoring
evidently==0.4.0

# Testing
pytest==7.4.1
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
joblib==1.3.1
tqdm==4.66.1
requests==2.31.0

# Production
gunicorn==21.2.0
python-multipart==0.0.6
